[Bugfix] UEA2: error when length is not a multiple of 4
This commit fixed the encryption error occurred when
the byte length is not a multiple of 4. Specifically speaking,
when the byte length is not a multiple of 4, the remaining
bytes are not processed at all.

In addition, this commit fixed the over-allocation bug
of keystream. The number of keys generated should be
the input's word length rather than it's byte length.